+++ src/Translator/Translator.php
@@ -1,9 +1,11 @@
 <?php
 
-namespace Yiisoft\I18n;
+namespace Yiisoft\I18n\Translator;
 
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Yiisoft\I18n\Event\MissingTranslationEvent;
+use Yiisoft\I18n\TranslationsLoaderInterface;
+use Yiisoft\I18n\TranslatorInterface;
 
 class Translator implements TranslatorInterface
 {
